requirements.txt
streamlit==1.32.0
pandas==2.2.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
bcrypt==4.1.2
python-jose==3.3.0  # for JWT tokens
passlib==1.7.4
python-dateutil==2.8.2
